Moved snowflake package out of collector and main app up in the new root.
