+++ src/main/java/com/robertozagni/SPYTM/data/collector/downloader/StockDataDownloaderRunner.java
@@ -4,7 +4,7 @@ import com.robertozagni.SPYTM.data.collector.downloader.alphavantage.AlphaVantag
 import com.robertozagni.SPYTM.data.collector.model.QuoteType;
 import com.robertozagni.SPYTM.data.collector.model.QuoteProvider;
 import com.robertozagni.SPYTM.data.collector.model.TimeSerie;
-import com.robertozagni.SPYTM.data.collector.datalake.service.SnowflakeStorageService;
+import com.robertozagni.SPYTM.data.datalake.service.SnowflakeStorageService;
 import com.robertozagni.SPYTM.data.collector.service.TimeSerieStorageService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
